# src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "oscp-spatial-content-discovery",
    "version": "1.0.0",
    "description": "OSCP Spatial Content Discovery (Open AR Cloud). Successful record payloads use the vendor media type `application/vnd.oscp+json; version=1.0`.",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Local dev"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "operationId": "getHealth",
        "tags": ["Health"],
        "responses": {
          "200": {
            "$ref": "#/components/responses/HealthOk"
          }
        }
      }
    },
    "/topics": {
      "get": {
        "summary": "List content topics served by this SCD instance",
        "operationId": "listTopics",
        "tags": ["Health"],
        "responses": {
          "200": {
            "description": "Lowercase topic names from this instance's TOPICS configuration",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "type": "string" },
                  "example": ["entertainment", "history", "transit"]
                }
              }
            }
          }
        }
      }
    },
    "/scrs/{topic}": {
      "get": {
        "summary": "Get SCRs covering an H3 index (and optional filters)",
        "operationId": "findScrsByHex",
        "tags": ["SCR"],
        "parameters": [
          { "$ref": "#/components/parameters/TopicPath" },
          { "$ref": "#/components/parameters/H3IndexQuery" },
          { "$ref": "#/components/parameters/PlacekeyQuery" },
          { "$ref": "#/components/parameters/KeywordsQuery" }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/OscpScrArray"
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        }
      },
      "post": {
        "summary": "Create a new SCR",
        "operationId": "createScr",
        "tags": ["SCR"],
        "parameters": [{ "$ref": "#/components/parameters/TopicPath" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ScrDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created; response body is the new record id",
            "content": {
              "text/plain": {
                "schema": { "type": "string" }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "security": [{ "bearerAuth": ["create:scrs"] }]
      }
    },
    "/scrs/{topic}/{id}": {
      "get": {
        "summary": "Get SCR by id",
        "operationId": "getScrById",
        "tags": ["SCR"],
        "parameters": [
          { "$ref": "#/components/parameters/TopicPath" },
          { "$ref": "#/components/parameters/IdPath" }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/OscpScr"
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        }
      },
      "put": {
        "summary": "Update SCR",
        "operationId": "updateScr",
        "tags": ["SCR"],
        "parameters": [
          { "$ref": "#/components/parameters/TopicPath" },
          { "$ref": "#/components/parameters/IdPath" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ScrDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated"
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "security": [{ "bearerAuth": ["update:scrs"] }]
      },
      "delete": {
        "summary": "Delete SCR",
        "operationId": "deleteScr",
        "tags": ["SCR"],
        "parameters": [
          { "$ref": "#/components/parameters/TopicPath" },
          { "$ref": "#/components/parameters/IdPath" }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "security": [{ "bearerAuth": ["delete:scrs"] }]
      }
    },
    "/tenant/scrs/{topic}": {
      "get": {
        "summary": "List SCRs for the authenticated tenant",
        "operationId": "listTenantScrs",
        "tags": ["SCR"],
        "parameters": [{ "$ref": "#/components/parameters/TopicPath" }],
        "responses": {
          "200": {
            "$ref": "#/components/responses/OscpScrArray"
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "security": [{ "bearerAuth": ["read:scrs"] }]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Auth0 JWT. When AUTH_REQUIRED=false (local/dev), protected routes accept requests without a token."
      }
    },
    "parameters": {
      "TopicPath": {
        "name": "topic",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Content topic (lowercased by the service)"
      },
      "IdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Platform-generated SCR id"
      },
      "H3IndexQuery": {
        "name": "h3Index",
        "in": "query",
        "required": false,
        "schema": { "type": "string" },
        "description": "H3 cell index for geographic search"
      },
      "PlacekeyQuery": {
        "name": "placekey",
        "in": "query",
        "required": false,
        "schema": { "type": "string" },
        "description": "Placekey (alternative to h3Index)"
      },
      "KeywordsQuery": {
        "name": "keywords",
        "in": "query",
        "required": false,
        "schema": { "type": "string" },
        "description": "Comma-separated keywords"
      }
    },
    "responses": {
      "HealthOk": {
        "description": "Service is healthy",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Health" }
          }
        }
      },
      "OscpScr": {
        "description": "OK",
        "content": {
          "application/vnd.oscp+json; version=1.0": {
            "schema": { "$ref": "#/components/schemas/Scr" }
          }
        }
      },
      "OscpScrArray": {
        "description": "OK",
        "content": {
          "application/vnd.oscp+json; version=1.0": {
            "schema": {
              "type": "array",
              "items": { "$ref": "#/components/schemas/Scr" }
            }
          }
        }
      },
      "Error400": {
        "description": "Bad request (validation, invalid topic, invalid h3Index, etc.)",
        "content": {
          "text/plain": {
            "schema": { "type": "string" }
          }
        }
      },
      "Error403": {
        "description": "Forbidden (invalid tenant / JWT claim)",
        "content": {
          "text/plain": {
            "schema": { "type": "string" }
          }
        }
      },
      "Error404": {
        "description": "Record not found",
        "content": {
          "text/plain": {
            "schema": { "type": "string" }
          }
        }
      },
      "Error500": {
        "description": "Internal server error",
        "content": {
          "text/plain": {
            "schema": { "type": "string" }
          }
        }
      }
    },
    "schemas": {
      "Health": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": {
            "type": "string",
            "example": "ok"
          }
        }
      },
      "Position": {
        "type": "object",
        "required": ["lon", "lat", "h"],
        "properties": {
          "lon": { "type": "number" },
          "lat": { "type": "number" },
          "h": { "type": "number" }
        }
      },
      "Quaternion": {
        "type": "object",
        "required": ["x", "y", "z", "w"],
        "properties": {
          "x": { "type": "number" },
          "y": { "type": "number" },
          "z": { "type": "number" },
          "w": { "type": "number" }
        }
      },
      "GeoPose": {
        "type": "object",
        "required": ["position", "quaternion"],
        "properties": {
          "position": { "$ref": "#/components/schemas/Position" },
          "quaternion": { "$ref": "#/components/schemas/Quaternion" }
        }
      },
      "Ref": {
        "type": "object",
        "required": ["contentType", "url"],
        "properties": {
          "contentType": { "type": "string" },
          "url": {
            "type": "string",
            "description": "Absolute http(s) URL, or a root-relative path in the client public folder (for example /media/pointclouds/cloud1.ply).",
            "pattern": "^(https?:\\/\\/[^\\s]+|\\/(?!\\/)[\\w\\-./%~]+)$"
          }
        }
      },
      "Def": {
        "type": "object",
        "required": ["type", "value"],
        "properties": {
          "type": { "type": "string" },
          "value": { "type": "string" }
        }
      },
      "Content": {
        "type": "object",
        "required": ["id", "type", "title", "geopose"],
        "properties": {
          "id": { "type": "string" },
          "type": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "keywords": {
            "type": "array",
            "items": { "type": "string" }
          },
          "placekey": { "type": "string" },
          "refs": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Ref" }
          },
          "geopose": { "$ref": "#/components/schemas/GeoPose" },
          "framedPose": {
            "type": "object",
            "description": "Optional opaque SpatialDDS FramedPose. Not validated by this geographic service."
          },
          "size": { "type": "number" },
          "bbox": { "type": "string" },
          "definitions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Def" }
          }
        }
      },
      "Scr": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "type": { "type": "string" },
          "content": { "$ref": "#/components/schemas/Content" },
          "tenant": { "type": "string" },
          "timestamp": { "type": "number" }
        }
      },
      "ScrDto": {
        "type": "object",
        "required": ["type", "content"],
        "properties": {
          "type": {
            "type": "string",
            "enum": ["scr"]
          },
          "content": { "$ref": "#/components/schemas/Content" }
        }
      }
    }
  }
}